src/main/java/com/rtg/variant/cnv/segment/AdjacentSegments.java
/*
 * Copyright (c) 2018. Real Time Genomics Limited.
 *
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are
 * met:
 *
 * 1. Redistributions of source code must retain the above copyright
 *    notice, this list of conditions and the following disclaimer.
 *
 * 2. Redistributions in binary form must reproduce the above copyright
 *    notice, this list of conditions and the following disclaimer in the
 *    documentation and/or other materials provided with the
 *    distribution.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
 * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
 * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
 * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */

package com.rtg.variant.cnv.segment;

/**
 * Hold a pair of adjacent segments and their score.
 */
class AdjacentSegments {
  private final double mScore;
  private final Segment mFirst;
  private final Segment mSecond;

  /**
   * @param score the score assigned to this pairing of segments
   * @param first the first of two segments. Usually the one with lowest starting position.
   * @param second the second of two segments. Usually the higher starting position.
   */
  AdjacentSegments(double score, Segment first, Segment second) {
    this.mScore = score;
    this.mFirst = first;
    this.mSecond = second;
  }

  /** @return the score assigned to these segments */
  public double getScore() {
    return mScore;
  }

  /** @return the first of the two segments */
  public Segment getFirst() {
    return mFirst;
  }

  /** @return the second of the two segments */
  public Segment getSecond() {
    return mSecond;
  }
}
